Fixed test errors
