+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/KeyValueGroupedDatasetE2ETestSuite.scala
@@ -20,9 +20,6 @@ import java.sql.Timestamp
 import java.util.Arrays
 
 import io.grpc.StatusRuntimeException
-import org.scalatest.concurrent.Eventually.eventually
-import org.scalatest.concurrent.Futures.timeout
-import org.scalatest.time.SpanSugar._
 
 import org.apache.spark.sql.catalyst.streaming.InternalOutputModes.Append
 import org.apache.spark.sql.connect.client.util.QueryTest
@@ -41,20 +38,6 @@ class KeyValueGroupedDatasetE2ETestSuite extends QueryTest with SQLHelper {
 
   lazy val session: SparkSession = spark
   import session.implicits._
-
-  val flatMapGroupsWithStateSchema: StructType = StructType(
-    Array(StructField("id", StringType), StructField("timestamp", TimestampType)))
-
-  val flatMapGroupsWithStateData: Seq[ClickEvent] = Seq(
-    ClickEvent("a", new Timestamp(12345)),
-    ClickEvent("a", new Timestamp(12346)),
-    ClickEvent("a", new Timestamp(12347)),
-    ClickEvent("b", new Timestamp(12348)),
-    ClickEvent("b", new Timestamp(12349)),
-    ClickEvent("c", new Timestamp(12350)))
-
-  val flatMapGroupsWithStateInitialStateData: Seq[ClickState] =
-    Seq(ClickState("a", 2), ClickState("b", 1))
 
   test("mapGroups") {
     val session: SparkSession = spark
@@ -478,7 +461,16 @@ class KeyValueGroupedDatasetE2ETestSuite extends QueryTest with SQLHelper {
         if (state.exists) throw new IllegalArgumentException("state.exists should be false")
         Iterator(ClickState(key, values.size))
       }
-    val values = flatMapGroupsWithStateData
+
+    val session: SparkSession = spark
+    import session.implicits._
+    val values = Seq(
+      ClickEvent("a", new Timestamp(12345)),
+      ClickEvent("a", new Timestamp(12346)),
+      ClickEvent("a", new Timestamp(12347)),
+      ClickEvent("b", new Timestamp(12348)),
+      ClickEvent("b", new Timestamp(12349)),
+      ClickEvent("c", new Timestamp(12350)))
       .toDS()
       .groupByKey(_.id)
       .flatMapGroupsWithState(Append, GroupStateTimeout.NoTimeout)(stateFunc)
@@ -492,12 +484,20 @@ class KeyValueGroupedDatasetE2ETestSuite extends QueryTest with SQLHelper {
         val currState = state.getOption.getOrElse(ClickState(key, 0))
         Iterator(ClickState(key, currState.count + values.size))
       }
-    val initialState = flatMapGroupsWithStateInitialStateData
-      .toDS()
-      .groupByKey(_.id)
-      .mapValues(x => x)
 
-    val values = flatMapGroupsWithStateData
+    val session: SparkSession = spark
+    import session.implicits._
+
+    val initialStateDS = Seq(ClickState("a", 2), ClickState("b", 1)).toDS()
+    val initialState = initialStateDS.groupByKey(_.id).mapValues(x => x)
+
+    val values = Seq(
+      ClickEvent("a", new Timestamp(12345)),
+      ClickEvent("a", new Timestamp(12346)),
+      ClickEvent("a", new Timestamp(12347)),
+      ClickEvent("b", new Timestamp(12348)),
+      ClickEvent("b", new Timestamp(12349)),
+      ClickEvent("c", new Timestamp(12350)))
       .toDS()
       .groupByKey(_.id)
       .flatMapGroupsWithState(Append, GroupStateTimeout.NoTimeout, initialState)(stateFunc)
@@ -511,7 +511,16 @@ class KeyValueGroupedDatasetE2ETestSuite extends QueryTest with SQLHelper {
         if (state.exists) throw new IllegalArgumentException("state.exists should be false")
         ClickState(key, values.size)
       }
-    val values = flatMapGroupsWithStateData
+
+    val session: SparkSession = spark
+    import session.implicits._
+    val values = Seq(
+      ClickEvent("a", new Timestamp(12345)),
+      ClickEvent("a", new Timestamp(12346)),
+      ClickEvent("a", new Timestamp(12347)),
+      ClickEvent("b", new Timestamp(12348)),
+      ClickEvent("b", new Timestamp(12349)),
+      ClickEvent("c", new Timestamp(12350)))
       .toDS()
       .groupByKey(_.id)
       .mapGroupsWithState(GroupStateTimeout.NoTimeout)(stateFunc)
@@ -525,177 +534,24 @@ class KeyValueGroupedDatasetE2ETestSuite extends QueryTest with SQLHelper {
         val currState = state.getOption.getOrElse(ClickState(key, 0))
         ClickState(key, currState.count + values.size)
       }
-    val initialState = flatMapGroupsWithStateInitialStateData
-      .toDS()
-      .groupByKey(_.id)
-      .mapValues(x => x)
-    val values = flatMapGroupsWithStateData
+
+    val session: SparkSession = spark
+    import session.implicits._
+
+    val initialStateDS = Seq(ClickState("a", 2), ClickState("b", 1)).toDS()
+    val initialState = initialStateDS.groupByKey(_.id).mapValues(x => x)
+
+    val values = Seq(
+      ClickEvent("a", new Timestamp(12345)),
+      ClickEvent("a", new Timestamp(12346)),
+      ClickEvent("a", new Timestamp(12347)),
+      ClickEvent("b", new Timestamp(12348)),
+      ClickEvent("b", new Timestamp(12349)),
+      ClickEvent("c", new Timestamp(12350)))
       .toDS()
       .groupByKey(_.id)
       .mapGroupsWithState(GroupStateTimeout.NoTimeout, initialState)(stateFunc)
 
     checkDataset(values, ClickState("a", 5), ClickState("b", 3), ClickState("c", 1))
-  }
-
-  test("flatMapGroupsWithState - streaming") {
-    val stateFunc =
-      (key: String, values: Iterator[ClickEvent], state: GroupState[ClickState]) => {
-        if (state.exists) throw new IllegalArgumentException("state.exists should be false")
-        Iterator(ClickState(key, values.size))
-      }
-    spark.sql("DROP TABLE IF EXISTS my_sink")
-
-    withTempPath { dir =>
-      val path = dir.getCanonicalPath
-      flatMapGroupsWithStateData.toDS().write.parquet(path)
-      val q = spark.readStream
-        .schema(flatMapGroupsWithStateSchema)
-        .parquet(path)
-        .as[ClickEvent]
-        .groupByKey(_.id)
-        .flatMapGroupsWithState(Append, GroupStateTimeout.NoTimeout)(stateFunc)
-        .writeStream
-        .format("memory")
-        .queryName("my_sink")
-        .start()
-
-      try {
-        q.processAllAvailable()
-        eventually(timeout(30.seconds)) {
-          checkDataset(
-            spark.table("my_sink").toDF().as[ClickState],
-            ClickState("c", 1),
-            ClickState("b", 2),
-            ClickState("a", 3))
-        }
-      } finally {
-        q.stop()
-        spark.sql("DROP TABLE IF EXISTS my_sink")
-      }
-    }
-  }
-
-  test("flatMapGroupsWithState - streaming - with initial state") {
-    val stateFunc =
-      (key: String, values: Iterator[ClickEvent], state: GroupState[ClickState]) => {
-        val currState = state.getOption.getOrElse(ClickState(key, 0))
-        Iterator(ClickState(key, currState.count + values.size))
-      }
-    val initialState = flatMapGroupsWithStateInitialStateData
-      .toDS()
-      .groupByKey(_.id)
-      .mapValues(x => x)
-    spark.sql("DROP TABLE IF EXISTS my_sink")
-
-    withTempPath { dir =>
-      val path = dir.getCanonicalPath
-      flatMapGroupsWithStateData.toDS().write.parquet(path)
-      val q = spark.readStream
-        .schema(flatMapGroupsWithStateSchema)
-        .parquet(path)
-        .as[ClickEvent]
-        .groupByKey(_.id)
-        .flatMapGroupsWithState(Append, GroupStateTimeout.NoTimeout, initialState)(stateFunc)
-        .writeStream
-        .format("memory")
-        .queryName("my_sink")
-        .start()
-
-      try {
-        q.processAllAvailable()
-        eventually(timeout(30.seconds)) {
-          checkDataset(
-            spark.table("my_sink").toDF().as[ClickState],
-            ClickState("c", 1),
-            ClickState("b", 3),
-            ClickState("a", 5))
-        }
-      } finally {
-        q.stop()
-        spark.sql("DROP TABLE IF EXISTS my_sink")
-      }
-    }
-  }
-
-  test("mapGroupsWithState - streaming") {
-    val stateFunc =
-      (key: String, values: Iterator[ClickEvent], state: GroupState[ClickState]) => {
-        if (state.exists) throw new IllegalArgumentException("state.exists should be false")
-        ClickState(key, values.size)
-      }
-    spark.sql("DROP TABLE IF EXISTS my_sink")
-
-    withTempPath { dir =>
-      val path = dir.getCanonicalPath
-      flatMapGroupsWithStateData.toDS().write.parquet(path)
-      val q = spark.readStream
-        .schema(flatMapGroupsWithStateSchema)
-        .parquet(path)
-        .as[ClickEvent]
-        .groupByKey(_.id)
-        .mapGroupsWithState(GroupStateTimeout.NoTimeout)(stateFunc)
-        .writeStream
-        .format("memory")
-        .queryName("my_sink")
-        .outputMode("update")
-        .start()
-
-      try {
-        q.processAllAvailable()
-        eventually(timeout(30.seconds)) {
-          checkDataset(
-            spark.table("my_sink").toDF().as[ClickState],
-            ClickState("c", 1),
-            ClickState("b", 2),
-            ClickState("a", 3))
-        }
-      } finally {
-        q.stop()
-        spark.sql("DROP TABLE IF EXISTS my_sink")
-      }
-    }
-  }
-
-  test("mapGroupsWithState - streaming - with initial state") {
-    val stateFunc =
-      (key: String, values: Iterator[ClickEvent], state: GroupState[ClickState]) => {
-        val currState = state.getOption.getOrElse(ClickState(key, 0))
-        ClickState(key, currState.count + values.size)
-      }
-    val initialState = flatMapGroupsWithStateInitialStateData
-      .toDS()
-      .groupByKey(_.id)
-      .mapValues(x => x)
-    spark.sql("DROP TABLE IF EXISTS my_sink")
-
-    withTempPath { dir =>
-      val path = dir.getCanonicalPath
-      flatMapGroupsWithStateData.toDS().write.parquet(path)
-      val q = spark.readStream
-        .schema(flatMapGroupsWithStateSchema)
-        .parquet(path)
-        .as[ClickEvent]
-        .groupByKey(_.id)
-        .mapGroupsWithState(GroupStateTimeout.NoTimeout, initialState)(stateFunc)
-        .writeStream
-        .format("memory")
-        .queryName("my_sink")
-        .outputMode("update")
-        .start()
-
-      try {
-        q.processAllAvailable()
-        eventually(timeout(30.seconds)) {
-          checkDataset(
-            spark.table("my_sink").toDF().as[ClickState],
-            ClickState("c", 1),
-            ClickState("b", 3),
-            ClickState("a", 5))
-        }
-      } finally {
-        q.stop()
-        spark.sql("DROP TABLE IF EXISTS my_sink")
-      }
-    }
   }
 }
